requirements_3.10.txt
absl-py                      2.1.0      
astunparse                   1.6.3      
audioread                    3.0.1
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.4.0
contourpy                    1.3.0
cycler                       0.12.1
decorator                    5.1.1
filelock                     3.16.1
flatbuffers                  24.3.25
fonttools                    4.54.1
fsspec                       2024.10.0
gast                         0.6.0
google-pasta                 0.2.0
grpcio                       1.67.1
h5py                         3.12.1
idna                         3.10
inflect                      7.4.0
Jinja2                       3.1.4
joblib                       1.4.2
#keras                        3.6.0
kiwisolver                   1.4.7
lazy_loader                  0.4
libclang                     18.1.1
librosa                      0.9.2
llvmlite                     0.43.0
Markdown                     3.7
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
matplotlib                   3.9.2
mdurl                        0.1.2
ml-dtypes                    0.4.1
more-itertools               10.5.0
mpmath                       1.3.0
msgpack                      1.1.0
namex                        0.0.8
networkx                     3.4.2
numba                        0.60.0
numpy                        2.0.2
opt_einsum                   3.4.0
optree                       0.13.0
packaging                    24.1
pillow                       11.0.0
#pip                          24.3.1
platformdirs                 4.3.6
pooch                        1.8.2
protobuf                     5.28.3
pycparser                    2.22
Pygments                     2.18.0
pyparsing                    3.2.0
python-dateutil              2.9.0.post0
requests                     2.32.3
resampy                      0.4.3
rich                         13.9.4
scikit-learn                 1.5.2
scipy                        1.14.1
setuptools                   75.3.0
six                          1.16.0
soundfile                    0.12.1
soxr                         0.5.0.post1
sympy                        1.13.3
# tensorboard                  2.18.0
# tensorboard-data-server      0.7.2
# tensorflow-cpu               2.18.0
# tensorflow_intel             2.18.0
# tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.5.0
threadpoolctl                3.5.0
# torch                        2.4.1+cpu
# torchaudio                   2.4.1+cpu
# torchvision                  0.19.1+cpu
typeguard                    4.4.1
typing_extensions            4.12.2
Unidecode                    1.3.8
urllib3                      2.2.3
Werkzeug                     3.1.1
wheel                        0.44.0
wrapt                        1.16.0